+++ src/main/java/yesman/epicfight/skill/mover/PhantomAscentSkill.java
@@ -13,7 +13,7 @@ import yesman.epicfight.api.animation.LivingMotions;
 import yesman.epicfight.api.animation.types.StaticAnimation;
 import yesman.epicfight.api.client.camera.EpicFightCameraAPI;
 import yesman.epicfight.api.client.input.MovementDirection;
-import yesman.epicfight.api.client.input.action.EpicFightInputActions;
+import yesman.epicfight.api.client.input.action.EpicFightInputAction;
 import yesman.epicfight.api.client.input.handlers.InputManager;
 import yesman.epicfight.api.client.neoevent.MappedMovementInputUpdateEvent;
 import yesman.epicfight.api.neoevent.playerpatch.PlayerPatchEvent;
@@ -157,6 +157,6 @@ public class PhantomAscentSkill extends Skill {
 	}
 
     private static boolean isJumpActionPressed() {
-        return InputManager.isActionActive(EpicFightInputActions.JUMP);
+        return InputManager.isActionActive(EpicFightInputAction.JUMP);
     }
 }