+++ RELEASE_NOTES.md
@@ -30,6 +30,11 @@
   causing issues with its intended functionality.
 - Fixed the mining crosshair not to show in vanilla mode
 
+### For Devs
+
+- Rename the experimental enum `EpicFightInputActions` to `EpicFightInputAction` to follow Java naming
+  conventions. [#2194](https://github.com/Epic-Fight/epicfight/issues/2194)
+
 ## [21.13.5] - 2025-11-12
 
 ### Fixed